[TinyLang][Tests] Add gtest support for TinyLang.
Original book source does not contain sample code for testing, so I had
to add testing because make it easier to follow the samples of the book
and also allows to check if the code builds and works as intended. So I
added a new target called "tinylang-tests" to include at minimum, unit
tests for each class.
Added an initial test for the Token class that indirectly tests the
Basic/TokenKinds.h header and helper functions.
Finally, created a tinylanglib target that will contain the source files
and a separate tinylang target for the compiler driver, in that way, the
testing and compiler driver targets can share the compiler source
library.