[TinyLang] Initial implementation of Parser and Semantic Analysis
In this commit, I started the implementation of the Parser and Semantic
Analysis. I added an initial test in the file tests/parser.cpp to make
sure the Parser and Semantic analysis can, at minimum, compile an Empty
module with a name. Then I'll add more tests to check for each
declaration, statement and expression implemented in the Parser.